fixed integer calculation to be double

+++ src/main/scala/com/ap/games/cards/CardGame.scala
@@ -35,7 +35,7 @@ case class CardGame(override val initialState: CardState) extends Game[CardState
 
   override def reward(state: CardState): Double = {
     val lifeLost = state.hero.maxLife - state.hero.life
-    val variance = lifeLost / state.hero.maxLife
+    val variance = lifeLost / state.hero.maxLife.toDouble
     val score = Math.max(0, state.deadEnemies.values.map(_.maxLife).sum + state.deadEnemies.size - variance)
     score / state.maxReward
   }
